BISERVER-1683
Spoke with the MSC team (and verified) that they've removed the optimizations discussed. If they do use the optimizations, they keep a local copy of the gwt.xml with the optimizations and don't check it in.  I added a comment to the gwt.xml files stating that optimizations are not to be put in production files.