[PMD-417]
Reverted to 1.7 code where signatures of security-related methods referenced IAclHolder instead of Object.  The original motivation for using Object was to distinguish between an object being protected and an object that knows how to fetch the permission bits associated with an object. Eventually all references to IAclHolder would have been changed to a service that when given an Object, would return the permission bits.  That service has not (and may never) come to fruition so it was decided to revert to the IAclHolder references in the method signatures.