Stronger clean-up. Most of the reports in the platform have a preferred-output type set, but the value given is horribly wrong (even to the lax standards of the "outputType" property. We now try our very best to map these values to something valid issueing loads of warnings on our way through the code whenever we hit entries that are not valid according to the strict and sane rules implemented now. With some luck, this will create only a small confusion and prevents a huge disaster later down the road when we have to long-term maintain this system.
Also fixes PRD-2338, which was caused by switching from action-inputs to a data-row. The fix now undoes that change and
removes the manual formula-evaluation, as this is also done in the DefaultParameterValidator.

Added a some logging to the  components to give better feedback on what is happening. After all, in a sane system the 
debug-log should contain enough info to spot errors instead of having to attach a debugger to the process. How do we
attach a debugger to a customer's computer?