[PDB-626]
Fixes many layout / resizing bugs with this class. It used to be extremely unpredictable and would constantly mess up the column sizes. Also the headers and columns would go out of sync because of some weird padding and calculations.

This commit goes in pair with a subsequent one on the pentaho-xul-gwt project.