BISERVER-4059
The solution/path duality is not reflected in the API - solutionPath is the monolithic property that contains it all. We have to split it out manually.

Also: XActions should contain the url attribute like everyone else - this safes me from coding a "special case" just to support that type. And above all - once everything changes into a plugin, I do not have to touch my code.

Parameter-Service in HttpWebService needs to set a proper error code when something evil happens. I do hate it when I cant tell the status easily.