http://jira/browse/BISERVER-4321 As a BI practitioner I want to upload a zipped or tar-gz'ed CSV
I modified UploadFileServlet to look for a parameter called 'unzip'. 
If 'unzip' is true and the uploaded file has a MIME type of 'application/zip', or 'application/x-compressed', or 'application/tgz', or 'application/x-tar', or an extension of '.zip', '.gzip', '.gz', '.tgz', or '.tar' the uploaded file is unpacked.

If the 'mark_temporary' parameter is true the files will be extracted into the tmp directory. The extracted files will have an extension of '.tmp', and the original file extension will precede it, e.g. myfile.csv will be extracted to [tmp file GUID].csv.tmp

The folder structure within the uploaded file will not be reproduced.

Files starting with a '.' or '__MACOSX' will not be extracted

The UploadFileServlet will return a newline separated list of the filenames unpacked from the uploaded file.