BISERVER-5665 - Content linking does not work when the link contains an apostrophe
now substitute single quotes for escaped versions in javascript links. this only fixes JFreeChart charts, OpenFlashCharts use XStream (JsonHierarchicalStreamDriver in particular) to generate the JSON, another bug lies in there. it actually translates escaped characters (anything leading with a backslash) literally.

so in java to escape single quotes in a String we do this "Home\\'s" when we really want "Home\'s". however, XStream sees both backslashes as bytes and renders a character for them... They come out as "Home\\'s". in the case of this issue, it causes a javascript error.