ANALYZER-832 - Coding of PDF & Excel Export as well as Scheduling
Now getting the location of the wkhtmltoimage tool from pentaho.xml. if the setting is not found in the file, it will be assumed that the tool can be executed by name (so it would be on the path (windows) or in a bin directory (linux/mac)). the setting is: <html-to-image-tool>