[PDI-8147] - CubeSpinner configure step
Added a FieldsetPane dojo widget. This is the panel that Brett has been asking for for a while. It has a label with a horizontal line next to it. The panel also has the disable & enable capability.


Added supporting styles for the new FieldsetPane.

Added themable images for chart, settings, and visualize.